refactor(CommentPreprocessor): preprocessing package name changed due to existing package
